Check fileMapping.includeDirectories in filter
By default, filterFiles includes directories in the %files list when captured by a glob pattern.  For files in system directories like /etc/init.d, the RPM will fail to install because it doesn't own the directory (e.g. "conflicts with file from package chkconfig").

This change disables it by default.  Specifying the includeDirectories: true in the file mapping will override this, and list directories in the spec file.